Separated data actions in a datamodulename_query and datamodulename_insert series of files for ease of usage and implementation
