+++ datatier_actions/refdata_query.py
@@ -7,7 +7,25 @@ from common.platform_settings import build_platform_variables
 from common.platform_settings import build_platform_config
 from common.auditerror_mgmt import process_auditerror_details
 
-def query_datatierdata_general(platform_vars, platform_settings, sql_connection, table_name)->list:
+def query_refdata_vendors(sql_connection)->list:
+    core_component = "refdata_vendors"
+    try:
+        sql_cursor = sql_connection.cursor()
+        sql_cursor.execute("SELECT * FROM refdata_vendors")
+        vendor_dtls = sql_cursor.fetchall()
+        rec_count = sql_cursor.rowcount
+        process_auditerror_details(platform_vars, platform_settings, auditerror_type="audit",
+                                   component_name="refdata", operation_name="query_vendors",
+                                   start_datetime=start_datetime, end_datetime=datetime.now(),
+                                   transaction_count=rec_count, error_id="NA",
+                                   error_desc="NA", processed_objectname="NA", audit_details="NA")
+    except:
+        print("Exception in query: " + {core_component})
+    finally:
+        sql_cursor.close()
+        return vendor_dtls
+
+def query_refdata_general(platform_vars, platform_settings, sql_connection, table_name)->list:
     try:
         # Auditing
         start_datetime = datetime.now()
@@ -18,7 +36,7 @@ def query_datatierdata_general(platform_vars, platform_settings, sql_connection,
         data_dtls = sql_cursor.fetchall()
         rec_count = sql_cursor.rowcount
         process_auditerror_details(platform_vars, platform_settings, auditerror_type="audit",
-                                   component_name="datatier", operation_name="query_"+table_name,
+                                   component_name="refdata", operation_name="query_"+table_name,
                                    start_datetime=start_datetime, end_datetime=datetime.now(),
                                    transaction_count=rec_count, error_id="NA",
                                    error_desc="NA", processed_objectname="NA", audit_details="NA")
@@ -28,7 +46,7 @@ def query_datatierdata_general(platform_vars, platform_settings, sql_connection,
         sql_cursor.close()
         return data_dtls
 
-def query_datatierdata_general_activerecords(platform_vars, platform_settings, sql_connection, table_name)->list:
+def query_refdata_general_activerecords(platform_vars, platform_settings, sql_connection, table_name)->list:
     try:
         # Auditing
         start_datetime = datetime.now()
@@ -36,14 +54,10 @@ def query_datatierdata_general_activerecords(platform_vars, platform_settings, s
         sql_cursor = sql_connection.cursor()
         sql_query = "select * from "+table_name+" where status_id='Active' "
         sql_cursor.execute(sql_query)
+        data_dtls = sql_cursor.fetchall()
         rec_count = sql_cursor.rowcount
-        if (rec_count >= 100):
-            data_dtls = sql_cursor.fetchmany(100)
-        else:
-            data_dtls = sql_cursor.fetchall()
-
         process_auditerror_details(platform_vars, platform_settings, auditerror_type="audit",
-                                   component_name="datatier", operation_name="query_"+table_name+"_activerecs",
+                                   component_name="refdata", operation_name="query_"+table_name+"_activerecs",
                                    start_datetime=start_datetime, end_datetime=datetime.now(),
                                    transaction_count=rec_count, error_id="NA",
                                    error_desc="NA", processed_objectname="NA", audit_details="NA")
@@ -53,18 +67,6 @@ def query_datatierdata_general_activerecords(platform_vars, platform_settings, s
         sql_cursor.close()
         return data_dtls
 
-def insert_datatier_crawlers()->None:
-    print(f"Starting Insert/Upsert Operation")
-
-
-def insert_datatier_sdp_dataattributes()->None:
-    print(f"Starting Insert/Upsert Operation")
-
-def insert_datatier_sdp_datastructure()->None:
-    print(f"Starting Insert/Upsert Operation")
-
-def insert_datatier_tokens()->None:
-    print(f"Starting Insert/Upsert Operation")
 
 if __name__ == "__main__":
     start_datetime = datetime.now()
@@ -76,11 +78,16 @@ if __name__ == "__main__":
         postgres_sql_connection = connectors.postgresql.create_connection(platform_settings.platform_datatier);
         #query_refdata_statuses(sql_connection=postgres_sql_connection)
         #create list of all reference tables
-        # Omitted - datatier_sdp_dataattributes as it has over 1.58 million recs when loaded
-        datatier_tables = ["datatier_crawlers","datatier_sdp_dataattributes",
-                           "datatier_sdp_datastructures","datatier_tokens"]
-        for table_name in datatier_tables:
-            query_datatierdata_general(platform_vars=platform_vars, platform_settings=platform_settings,
-                                       sql_connection=postgres_sql_connection, table_name=table_name)
-            query_datatierdata_general_activerecords(platform_vars=platform_vars,platform_settings=platform_settings,
-                                       sql_connection=postgres_sql_connection, table_name=table_name)
+        refdata_tables = ["refdata_applications", "refdata_codesets", "refdata_dataattributes",
+                          "refdata_datastructures", "refdata_devicetypes", "refdata_industries",
+                          "refdata_industries_business","refdata_industrystds",
+                          "refdata_industrystds_eventtypes", "refdata_legalentities",
+                          "refdata_operationtypes", "refdata_organizations", "refdata_professiontypes",
+                          "refdata_rulesets", "refdata_sensitivityflags", "refdata_status",
+                          "refdata_terminologystds","refdata_timezones","refdata_usstates","refdata_vendors"]
+        for refdata_table in refdata_tables:
+            query_refdata_general(platform_vars=platform_vars,platform_settings=platform_settings,
+                                  sql_connection=postgres_sql_connection, table_name=refdata_table)
+            query_refdata_general_activerecords(platform_vars=platform_vars, platform_settings=platform_settings,
+                                  sql_connection=postgres_sql_connection, table_name=refdata_table)
+
